Move rnn.py to commons/parts/rnn.py
